Fix findings reported by the Whitaker Dylint suite
Prepare the codebase for adopting the Whitaker Dylint suite by fixing
every finding it reports:

- Split `map`, `dbsp_circuit::streams::behaviour`, `dbsp_sync::output`,
  and `test_utils::physics` into sibling submodules so no module
  exceeds the 400-line budget (`module_max_lines`).
- Add `//!` inner doc comments to `__macro_deps`, thirteen in-source
  test modules, and the `dbsp_test_support` helper module, and move
  the `#![cfg(feature = "test-support")]` attribute below the existing
  doc comment in the shared test-support files
  (`module_must_have_inner_docs`).
- Make test fixtures fallible instead of panicking: `new_circuit`
  (both the `test_utils` crate and the in-crate stream helper),
  `build_circuit`, `run_health_delta`, `fresh_state`, `make_state`,
  and `setup_app` now return `Result` and propagate errors; the test
  bodies unwrap, so a failure is the test verdict
  (`no_expect_outside_tests`).
- Replace the tick counter's panicking `checked_add().expect()` with
  `saturating_add` inside the generator closure.
- Convert `assert_health_delta` into a macro so panic locations point
  at the calling test and the fallible circuit run executes inside the
  test body.
- Build the rspec `Configuration` directly from its public fields
  instead of the fallible builder, removing the panicking
  `unwrap_or_else` fallback (`no_unwrap_or_else_panic`).
- Add a direct `dbsp` dependency to `test_utils` so its fixtures can
  name `dbsp::Error` in their return types.
