+++ src/lib.rs
@@ -23,8 +23,11 @@ pub use constants::*;
 
 #[doc(hidden)]
 pub mod __macro_deps {
-    // Public solely for cross-crate macro hygiene; do not depend on it directly.
-    // This module sits outside the semver-stable public API surface.
+    //! Re-exports consumed by this crate's exported macros.
+    //!
+    //! Public solely for cross-crate macro hygiene; do not depend on it
+    //! directly. This module sits outside the semver-stable public API
+    //! surface.
     pub use rkyv;
     pub use size_of;
 }
