+++ test_utils/Cargo.toml
@@ -6,6 +6,7 @@ publish = false
 
 [dependencies]
 syn = { version = "2.0.103", features = ["full"] }
+dbsp = "0.98"
 lille = { path = ".." }
 ordered-float = { workspace = true }
 bevy = { workspace = true, default-features = false }
